Move all router-related files to detail namespace.
