+++ include/telnetpp/detail/subnegotiation_router.hpp
@@ -1,14 +1,12 @@
 #ifndef TELNETPP_SUBNEGOTIATION_ROUTER_HPP_
 #define TELNETPP_SUBNEGOTIATION_ROUTER_HPP_
 
-#include "telnetpp/router.hpp"
+#include "telnetpp/detail/router.hpp"
 #include "telnetpp/subnegotiation.hpp"
 #include "telnetpp/element.hpp"
 #include <vector>
 
-namespace telnetpp {
-    
-namespace detail {
+namespace telnetpp { namespace detail {
 
 struct subnegotiation_router_key_from_message_policy
 {
@@ -18,30 +16,16 @@ struct subnegotiation_router_key_from_message_policy
     }
 };
 
-}
-
 class subnegotiation_router
   : public router<
         u8,
         subnegotiation,
-        std::vector<token>,
+        std::vector<telnetpp::token>,
         detail::subnegotiation_router_key_from_message_policy
     >
 {
 };
- 
-template <class SubnegotiableOption>
-void register_route_from_subnegotiation_to_option(
-    subnegotiation_router &route,
-    SubnegotiableOption   &option)
-{
-    route.register_route(option.option(), 
-        [&option](auto &&sub)
-        {
-            return option.subnegotiate(sub.content());
-        });
-}
 
-}
+}}
 
 #endif
