+++ src/server_option.cpp
@@ -1,7 +1,5 @@
 #include "telnetpp/server_option.hpp"
-#include "telnetpp/negotiation_router.hpp"
 #include "telnetpp/protocol.hpp"
-#include "telnetpp/subnegotiation_router.hpp"
 
 namespace telnetpp {
 
@@ -183,22 +181,6 @@ std::vector<telnetpp::token> server_option::handle_subnegotiation(
 {
     // By default, do nothing.
     return {};
-}
-
-// ==========================================================================
-// REGISTER_CLIENT_OPTION
-// ==========================================================================
-void register_server_option(
-    server_option &option, 
-    negotiation_router &neg_router, 
-    subnegotiation_router &sub_router)
-{
-    register_route_from_negotiation_to_option(
-        neg_router, telnetpp::do_, option);
-    register_route_from_negotiation_to_option(
-        neg_router, telnetpp::dont, option);
-    register_route_from_subnegotiation_to_option(
-        sub_router, option);
 }
 
 }
