+++ test/routing_visitor_test.cpp
@@ -1,7 +1,8 @@
-#include "telnetpp/routing_visitor.hpp"
+#include "telnetpp/detail/routing_visitor.hpp"
 #include "telnetpp/options/naws/client.hpp"
 #include "telnetpp/options/naws.hpp"
 #include "telnetpp/protocol.hpp"
+#include "telnetpp/detail/registration.hpp"
 #include "expect_elements.hpp"
 #include <cppunit/TestFixture.h>
 #include <cppunit/extensions/HelperMacros.h>
@@ -31,14 +32,14 @@ CPPUNIT_TEST_SUITE_REGISTRATION(routing_visitor_test);
 
 void routing_visitor_test::text_routes_to_text_function()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     
     std::string text;
     std::string expected_text = "expected text";
     
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         [&text](auto &&new_text) -> std::vector<telnetpp::token>
         {
             text = new_text;
@@ -56,11 +57,11 @@ void routing_visitor_test::text_routes_to_text_function()
 
 void routing_visitor_test::text_does_not_route_to_null_function()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         nullptr,
         cmd_router,
         neg_router,
@@ -72,9 +73,9 @@ void routing_visitor_test::text_does_not_route_to_null_function()
 
 void routing_visitor_test::command_routes_to_command_router()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     
     telnetpp::command cmd(0x00);
     telnetpp::command expected_command(telnetpp::ayt);
@@ -86,7 +87,7 @@ void routing_visitor_test::command_routes_to_command_router()
             return {};
         });
         
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         [](auto &&) -> std::vector<telnetpp::token>
         {
             return {};
@@ -103,9 +104,9 @@ void routing_visitor_test::command_routes_to_command_router()
 
 void routing_visitor_test::negotiation_routes_to_negotiation_router()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     telnetpp::options::naws::client client;
     client.activate();
     
@@ -117,9 +118,9 @@ void routing_visitor_test::negotiation_routes_to_negotiation_router()
             return {};
         });
     
-    telnetpp::register_client_option(client, neg_router, sub_router);
+    telnetpp::detail::register_client_option(client, neg_router, sub_router);
         
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         [](auto &&) -> std::vector<telnetpp::token>
         {
             return {};
@@ -140,9 +141,9 @@ void routing_visitor_test::negotiation_routes_to_negotiation_router()
 
 void routing_visitor_test::subnegotiation_routes_to_subnegotiation_router()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     telnetpp::options::naws::client client;
     client.activate();
     client.negotiate(telnetpp::will);
@@ -158,10 +159,10 @@ void routing_visitor_test::subnegotiation_routes_to_subnegotiation_router()
             return {};
         });
         
-    telnetpp::register_route_from_subnegotiation_to_option(
+    telnetpp::detail::register_route_from_subnegotiation_to_option(
         sub_router, client);
         
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         [](auto &&) -> std::vector<telnetpp::token>
         {
             return {};
@@ -186,12 +187,12 @@ void routing_visitor_test::subnegotiation_routes_to_subnegotiation_router()
 
 void routing_visitor_test::subnegotiation_accumulates_responses()
 {
-    telnetpp::command_router cmd_router;
-    telnetpp::negotiation_router neg_router;
-    telnetpp::subnegotiation_router sub_router;
+    telnetpp::detail::command_router cmd_router;
+    telnetpp::detail::negotiation_router neg_router;
+    telnetpp::detail::subnegotiation_router sub_router;
     telnetpp::options::naws::client client;
     
-    telnetpp::register_client_option(client, neg_router, sub_router);
+    telnetpp::detail::register_client_option(client, neg_router, sub_router);
     
     client.on_window_size_changed.connect(
         [&client](auto &&, auto &&)
@@ -199,7 +200,7 @@ void routing_visitor_test::subnegotiation_accumulates_responses()
             return client.deactivate();
         });
     
-    telnetpp::routing_visitor visitor(
+    telnetpp::detail::routing_visitor visitor(
         [](auto &&) -> std::vector<telnetpp::token>
         {
             return {};
